refactor/pithos: Refactor structs, fixed prelimenary compile errors
