+++ crates/pithos_lib/Cargo.toml
@@ -36,6 +36,7 @@ thiserror = "1.0.56"
 crypto_kx = "0.2.1"
 base64 = "0.21.7"
 scrypt = { version = "0.11.0" }
+borsh = {version = "1.3.1", features = ["std", "derive"]}
 
 [dev-dependencies]
 criterion = { version = "0.5.1", features = ["html_reports", "async_tokio"] }
