+++ crates/pithos_lib/src/helpers/mod.rs
@@ -1,2 +1,3 @@
 pub mod flag_helpers;
 pub mod footer_parser;
+pub mod structs;
