+++ crates/pithos_lib/src/helpers/structs.rs
@@ -1,0 +1,56 @@
+#[derive(Debug, PartialEq)]
+pub enum ProbeResult {
+    Unknown,
+    Compression,
+    NoCompression,
+}
+
+#[derive(Eq, PartialEq, PartialOrd, Ord, Clone, Copy, Debug)]
+pub struct Range {
+    pub from: u64,
+    pub to: u64,
+}
+
+#[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Debug, Default)]
+pub struct FileContext {
+    // FileName
+    pub file_name: String,
+    // Input size
+    pub input_size: u64,
+    // Filesize
+    pub file_size: u64,
+    // FileSubpath without filename
+    pub file_path: Option<String>,
+    // UserId
+    pub uid: Option<u64>,
+    // GroupId
+    pub gid: Option<u64>,
+    // Octal like mode
+    pub mode: Option<u32>,
+    // Created at
+    pub mtime: Option<u64>,
+    // Should this file be skipped by decompressors
+    pub compression: bool,
+    // Encryption key
+    pub encryption_key: Option<Vec<u8>>,
+    // Owners pubkey
+    pub owners_pubkey: Option<[u8; 32]>,
+    // Is this file a directory
+    pub is_dir: bool,
+    // Is this file a symlink
+    pub is_symlink: bool,
+    // Expected SHA1 hash
+    pub expected_sha1: Option<String>,
+    // Expected MD5 hash
+    pub expected_md5: Option<String>,
+}
+
+impl FileContext {
+    #[tracing::instrument(level = "trace", skip(self))]
+    pub fn get_path(&self) -> String {
+        match &self.file_path {
+            Some(p) => p.clone() + "/" + &self.file_name,
+            None => self.file_name.clone(),
+        }
+    }
+}
