+++ crates/pithos_lib/src/pithos/mod.rs
@@ -1,0 +1,3 @@
+pub mod pithosreader;
+pub mod pithoswriter;
+pub mod structs;
