+++ crates/pithos_lib/src/transformers/filter.rs
@@ -1,4 +1,4 @@
-use crate::helpers::footer_parser::Range;
+use crate::helpers::structs::Range;
 use crate::notifications::{Message, Notifier};
 use crate::transformer::{Transformer, TransformerType};
 use anyhow::{anyhow, Result};
